move current_time_millis
